Create data_analytics_machine_learning
A project completed during my MBA program at the University of California, Irvine for a Program & Analytics course. The underlying question to be answered through the respective analysis is: which variables, if any, predict the level of absenteeism in the workplace. In-scope variables for analysis are those relative to a person's lifestyle such as body health (e.g. weight), education level, and even smoking habits.

The benefit of the analysis is directly for companies -- to better understand the elements that contribute towards absenteeism in its employees,